personal/strategy/poc/voice-multispeaker/pyproject.toml
[project]
name = "voice-multispeaker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-texttospeech>=2.31.0",
    "google-auth>=2.0.0",
]

[tool.uv]
package = false
